backend/requirements.txt
# FastAPI core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9

# Pydantic / settings
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Object storage (S3-compatible)
boto3==1.35.0
botocore==1.35.0

# AI
anthropic==0.34.0

# Task queue
celery==5.4.0
redis==5.1.0

# Utilities
httpx==0.27.2
python-slugify==8.0.4
